requirements.txt
absl-py==1.4.0
accelerate==0.33.0
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
Pillow==10.1.0
protobuf==4.25.4
psutil==6.0.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
setuptools==70.2.0
six==1.16.0
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.5.0.dev20240809
torchaudio==2.4.0.dev20240809
torchvision==0.20.0.dev20240809
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
